Additional sanity checks on AMF packets.
AMF decided to allow negative length strings and arrays.  If the client
tries these shenanigans, just treat it as a 0 length object instead of
believing them.

Add a few checks to make sure we aren't reading past the end of the packet
if it ends too early.

Also don't let a client claim it's going to send a 2 billion arrays to
trick us into allocation 2 billion items to store them in.