Codec sanitization
ScreenVideo allocated more memory than necessary per block, and could run past EOF if the block was malformed.

ScreenVideo2 had the same issues as ScreenVideo, but calculated blockHeight and blockWidth incorrectly due to missing ().

ExtendedAudio did not rewind the pointer if it decided it couldn't handle the audio, and could cause an NPE on an unknown or invalid fourcc value.

AV1Video always thought data passed to it was AV1 data rather than checking the fourcc value.

MPEG1Video had the opposite problem, it looked to see if the data was AVC video data then tried to parse it as though it were MPEG1, and threw away the first configuration packet because decoderConfiguration wasn't initialized on the first pass.